fix: add workaround for not being able to `nc` to MongoDB port
Modify `wait-for.sh` script to try both nc and netstat (netstat only
when host is `localhost` or `127.0.0.1`).

For some reason, MongoDB container running in Alpine Linux in qemu
cannot be reached by `nc` from Alpine container in the same machine,
in the same pod. Maybe its' because MongoDB is based on Ubuntu?
Anyway, netstat sees port being listened to, so its enough in this case.