fix: handling of NodeBB in "cluster" mode
APP_USE_CLUSTER can be set to 2, 3, or more to run that many
instances of NodeBB.

If APP_USE_PORT is specified, it will be used as a "first" port.
So first NodeBB instance will be accessible through that port.
Second one will be accessible through APP_USE_PORT+1 port, etc...

Once we have APP_ADD_NGINX done, APP_USE_PORT will be kept as is,
and NGINX will be used as load balancer to internal NodeBB's ports.