feat: support NOBBIC_START_AUTOBUILD variable with "yes" or "no" values
Defaults to "yes", but OpenRC and SystemD service templates
set it to "no".

When not "yes", nobbic will not try to build pod from the `start` action.
It will simply exit with error, if pod does not exists.

This fixes problem reported by @nhlpl: when service was not removed,
after pod was removed, reboot caused building another pod with the same
name (but almost empty, and failing, because NodeBB needs database).
That could prevent another service from running correctly because of
using the same default port number(s).