fix: remove hard-coded port in `onbb_wait_until_db_ready()`
It was not only hard-coded, but misspelled too!
Everything was working fine, because it was used only with a last resort
try to check if something is listening on it, and grep was used so it
kinda was catching the right port (unless something else was listening
on 2701, in which case it would assume DB is ready, even if it was not).