fix: when checking availability of port, ignore whitespace
There's no need for `u` flag, as we do not read diff results, and it's
not available in BusyBox (Alpine) anyway.

But `w` flag is supported and allows to ignore stuff like "no newline
at the end", which we do not care about, as long as test string is
returned by test server.